implement first hardware timer

+++ src/hardware_abstraction/timer/i_timer.hpp
@@ -9,7 +9,7 @@ class ITimer
    public:
     ~ITimer() = default;
     virtual std::uint64_t get() = 0;
-    virtual void setAlarm(std::uint64_t alarm_value) = 0;
+    virtual void setAlarm(std::uint32_t alarm_value) = 0;
 };
 }  // namespace hardware_layer
